[IMP] Re-structure the code to generate SEPA Creditor Identifier
Don't set SchmeNm/Prtry to "SEPA" when the payment order is not SEPA.
